+++ dex-it-common/src/main/scala/com/wavesplatform/dex/it/api/dex/HasDex.scala
@@ -1,17 +1,20 @@
 package com.wavesplatform.dex.it.api.dex
 
+import java.util.Properties
 import java.util.concurrent.ThreadLocalRandom
 
 import cats.Functor
-import cats.instances.FutureInstances
 import com.typesafe.config.{Config, ConfigFactory}
 import com.wavesplatform.dex.it.api.BaseContainersKit
 import com.wavesplatform.dex.it.docker.base.info.DexContainerInfo
 import com.wavesplatform.dex.it.docker.base.{BaseContainer, DexContainer}
 import com.wavesplatform.dex.it.fp.CanExtract
 import mouse.any._
+import org.apache.kafka.clients.admin.{AdminClient, NewTopic}
 
-trait HasDex extends FutureInstances { self: BaseContainersKit =>
+import scala.collection.JavaConverters._
+
+trait HasDex { self: BaseContainersKit =>
 
   protected implicit def toDexExplicitGetOps[F[_]: CanExtract: Functor](self: DexApi[F]): DexApiOps.ExplicitGetDexApiOps[F] = {
     new DexApiOps.ExplicitGetDexApiOps[F](self)
@@ -38,4 +41,25 @@ trait HasDex extends FutureInstances { self: BaseContainersKit =>
   }
 
   lazy val dex1: DexContainer = createDex("dex-1")
+
+  protected def createKafkaTopic(name: String): Unit = kafkaServer.foreach { server =>
+    val properties = new Properties()
+    properties.putAll(
+      Map(
+        "bootstrap.servers"  -> server,
+        "group.id"           -> s"create-$name",
+        "key.deserializer"   -> "org.apache.kafka.common.serialization.StringDeserializer",
+        "value.deserializer" -> "org.apache.kafka.common.serialization.StringDeserializer"
+      ).asJava)
+
+    val adminClient = AdminClient.create(properties)
+    try {
+      val newTopic = new NewTopic(name, 1, 1.toShort)
+      adminClient.createTopics(java.util.Collections.singletonList(newTopic))
+    } finally {
+      adminClient.close()
+    }
+  }
+
+  protected def kafkaServer: Option[String] = Option(System.getenv("KAFKA_SERVER"))
 }
