+++ dex-it-common/src/main/scala/com/wavesplatform/dex/it/assets/DoubleOps.scala
@@ -4,7 +4,8 @@ import com.wavesplatform.dex.model.MatcherModel.Normalization
 
 trait DoubleOps {
   implicit final class DoubleOpsImplicits(val value: Double) {
-    val waves: Long = Normalization.normalizeAmountAndFee(value, 8)
+    val waves, eth, btc: Long = Normalization.normalizeAmountAndFee(value, 8)
+    val usd: Long             = Normalization.normalizePrice(value, 8, 2)
   }
 }
