+++ dex/src/main/resources/application.conf
@@ -299,7 +299,7 @@ waves.dex {
 
       # Consumer-related settings
       consumer {
-        # The consumer is waiting new messages in this duration. After this it retries. So there is a loop.
+        # The consumer is waiting for new messages in this duration. After this it retries. So there is a loop.
         fetch-max-duration = 30ms
 
         # A maximum number of polled messages in a buffer
@@ -316,7 +316,7 @@ waves.dex {
           group.id = ${waves.dex.events-queue.kafka.group}
 
           max.poll {
-            # Max interval between two polls. If it exceeded. the consumer is considered failed
+            # Max interval between two polls. If it exceeded, the consumer is considered failed
             interval.ms = 11000
 
             # The maximum number of records returned in a single call
